Fix for raspberry pi - add fake audio support.

+++ src/visualizer/power_circles.rs
@@ -48,7 +48,7 @@ impl PowerCirclesVisualizer {
                 let mut max_length = mem::uninitialized();
                 gl_try!(gl; gl.GetProgramiv(program, gl::INFO_LOG_LENGTH, &mut max_length));
 
-                let mut info_log = vec![0 as i8; max_length as usize];
+                let mut info_log = vec![0; max_length as usize];
                 gl_try!(gl; gl.GetProgramInfoLog(program, max_length, &mut max_length, info_log.as_mut_ptr()));
 
                 for info_char in info_log.iter() {
